requirements.txt
websockets
aiomqtt
python-dotenv